First attempt at fixing platform incompatibility issues (#5)
Because during the build process the `build-until` value is set to `201`, the plugin does not work with the build of `2020.2`.

Now the `build-until` restriction is jut removed, as I do not want to manually increase it every time that a new platform update is released. Most of the time it should just work anyway.